Extend and refactor the REST interface
The REST interface is extended with support for the "favorite" and "rank
history" operations, which allow the user to specify certain games as
favorites and obtain the rank history of a game. The old methods were
also refactored to use appId instead of gameId as the identifier, since
gameId seems to be undefined for several titles. The parameter orders
were also modified to make them more consistent between the different
methods.